Make Car a JPA entity mapped to the cars table
Complete the entity step of the Postgres migration (Milestone 2,
Part 2): the Car domain class now maps to the Flyway-managed cars
table instead of being a plain JSON-backed object.

- Annotate Car as @Entity/@Table(name="cars") with @Id +
  @GeneratedValue(IDENTITY), @Column constraints, and
  @Enumerated(STRING) for category; add the protected no-arg
  constructor Hibernate requires.
- Let the database own identity: id changes from int to Long and is
  null until a row is persisted. Propagate the type through
  Collection.nextID (now null-safe), JsonReader, and
  CarCollectionManager.
- Add @EntityScan("model") so Hibernate discovers Car, which lives
  outside the app package.
- Update TestCar/JsonWriterTest for the null id, and fix the
  backwards toJson assertion this surfaced.

+++ src/main/java/persistence/JsonReader.java
@@ -61,7 +61,7 @@ public class JsonReader {
     // MODIFIES: c
     // EFFECTS: parses car from JSON object and adds it to collection
     private void addCar(Collection c, JSONObject jsonObject) {
-        int id = jsonObject.getInt("id");
+        Long id = jsonObject.getLong("id");
         int year = jsonObject.getInt("year");
         String make = jsonObject.getString("make");
         String model = jsonObject.getString("model");
